# perf: use a reusable bytes.Buffer to speed up MutableTree.String