--- a/with_gcc_test.go
+++ b/with_gcc_test.go
@@ -1,3 +1,4 @@
+//go:build gcc
 // +build gcc
 
 // This file exists because some of the DBs e.g CLevelDB