@robert-zaremba 2021-12-16
Probably we could optimize it even further with `strings.Join`.